+++ packages/instrumentation-pg/test/pg-pool.test.ts
@@ -41,21 +41,28 @@ import * as pgPool from 'pg-pool';
 import { AttributeNames } from '../src/enums/AttributeNames';
 import { TimedEvent } from './types';
 import {
-  DBSYSTEMVALUES_POSTGRESQL,
-  SEMATTRS_DB_SYSTEM,
-  SEMATTRS_DB_NAME,
-  SEMATTRS_NET_PEER_NAME,
-  SEMATTRS_DB_CONNECTION_STRING,
-  SEMATTRS_NET_PEER_PORT,
-  SEMATTRS_DB_USER,
-  SEMATTRS_DB_STATEMENT,
+  METRIC_DB_CLIENT_OPERATION_DURATION,
+  DB_SYSTEM_NAME_VALUE_POSTGRESQL,
 } from '@opentelemetry/semantic-conventions';
 import {
   ATTR_DB_CLIENT_CONNECTION_STATE,
   METRIC_DB_CLIENT_CONNECTION_COUNT,
   METRIC_DB_CLIENT_CONNECTION_PENDING_REQUESTS,
-  METRIC_DB_CLIENT_OPERATION_DURATION,
+  ATTR_DB_SYSTEM,
+  ATTR_DB_NAME,
+  ATTR_DB_USER,
+  DB_SYSTEM_VALUE_POSTGRESQL,
+  ATTR_DB_CONNECTION_STRING,
+  ATTR_NET_PEER_PORT,
+  ATTR_NET_PEER_NAME,
+  ATTR_DB_STATEMENT,
 } from '../src/semconv';
+import {
+  ATTR_DB_QUERY_TEXT,
+  ATTR_DB_SYSTEM_NAME,
+  ATTR_SERVER_ADDRESS,
+  ATTR_SERVER_PORT,
+} from '@opentelemetry/semantic-conventions';
 
 const memoryExporter = new InMemorySpanExporter();
 
@@ -63,6 +70,7 @@ const CONFIG = {
   user: process.env.POSTGRES_USER || 'postgres',
   password: process.env.POSTGRES_PASSWORD || 'postgres',
   database: process.env.POSTGRES_DB || 'postgres',
+  namespace: 'namespace',
   host: process.env.POSTGRES_HOST || 'localhost',
   port: process.env.POSTGRES_PORT
     ? parseInt(process.env.POSTGRES_PORT, 10)
@@ -72,23 +80,29 @@ const CONFIG = {
 };
 
 const DEFAULT_PGPOOL_ATTRIBUTES = {
-  [SEMATTRS_DB_SYSTEM]: DBSYSTEMVALUES_POSTGRESQL,
-  [SEMATTRS_DB_NAME]: CONFIG.database,
-  [SEMATTRS_NET_PEER_NAME]: CONFIG.host,
-  [SEMATTRS_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
-  [SEMATTRS_NET_PEER_PORT]: CONFIG.port,
-  [SEMATTRS_DB_USER]: CONFIG.user,
+  [ATTR_DB_SYSTEM]: DB_SYSTEM_VALUE_POSTGRESQL,
+  [ATTR_DB_NAME]: CONFIG.database,
+  [ATTR_NET_PEER_NAME]: CONFIG.host,
+  [ATTR_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
+  [ATTR_NET_PEER_PORT]: CONFIG.port,
+  [ATTR_DB_USER]: CONFIG.user,
   [AttributeNames.MAX_CLIENT]: CONFIG.maxClient,
   [AttributeNames.IDLE_TIMEOUT_MILLIS]: CONFIG.idleTimeoutMillis,
 };
 
 const DEFAULT_PG_ATTRIBUTES = {
-  [SEMATTRS_DB_SYSTEM]: DBSYSTEMVALUES_POSTGRESQL,
-  [SEMATTRS_DB_NAME]: CONFIG.database,
-  [SEMATTRS_NET_PEER_NAME]: CONFIG.host,
-  [SEMATTRS_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
-  [SEMATTRS_NET_PEER_PORT]: CONFIG.port,
-  [SEMATTRS_DB_USER]: CONFIG.user,
+  [ATTR_DB_SYSTEM]: DB_SYSTEM_VALUE_POSTGRESQL,
+  [ATTR_DB_NAME]: CONFIG.database,
+  [ATTR_NET_PEER_NAME]: CONFIG.host,
+  [ATTR_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
+  [ATTR_NET_PEER_PORT]: CONFIG.port,
+  [ATTR_DB_USER]: CONFIG.user,
+};
+
+const STABLE_PG_QUERY_ATTRIBUTES = {
+  [ATTR_DB_SYSTEM_NAME]: DB_SYSTEM_NAME_VALUE_POSTGRESQL,
+  [ATTR_SERVER_ADDRESS]: CONFIG.host,
+  [ATTR_SERVER_PORT]: CONFIG.port,
 };
 
 const unsetStatus: SpanStatus = {
@@ -127,9 +141,8 @@ describe('pg-pool', () => {
     spanProcessors: [new SimpleSpanProcessor(memoryExporter)],
   });
 
-  const testPostgres = process.env.RUN_POSTGRES_TESTS; // For CI: assumes local postgres db is already available
-  const testPostgresLocally = process.env.RUN_POSTGRES_TESTS_LOCAL; // For local: spins up local postgres db via docker
-  const shouldTest = testPostgres || testPostgresLocally; // Skips these tests if false (default)
+  const testPostgres = process.env.RUN_POSTGRES_TESTS;
+  const shouldTest = testPostgres; // Skips these tests if false (default)
 
   before(function () {
     const skip = () => {
@@ -143,10 +156,6 @@ describe('pg-pool', () => {
       skip();
     }
 
-    if (testPostgresLocally) {
-      testUtils.startDocker('postgres');
-    }
-
     instrumentation = new PgInstrumentation();
 
     contextManager = new AsyncLocalStorageContextManager().enable();
@@ -158,10 +167,6 @@ describe('pg-pool', () => {
   });
 
   after(done => {
-    if (testPostgresLocally) {
-      testUtils.cleanUpDocker('postgres');
-    }
-
     pool.end(() => {
       done();
     });
@@ -196,7 +201,7 @@ describe('pg-pool', () => {
       };
       const pgAttributes = {
         ...DEFAULT_PG_ATTRIBUTES,
-        [SEMATTRS_DB_STATEMENT]: 'SELECT NOW()',
+        [ATTR_DB_STATEMENT]: 'SELECT NOW()',
       };
       const events: TimedEvent[] = [];
       const span = provider.getTracer('test-pg-pool').startSpan('test span');
@@ -230,12 +235,12 @@ describe('pg-pool', () => {
       });
 
       const expectedAttributes = {
-        [SEMATTRS_DB_SYSTEM]: DBSYSTEMVALUES_POSTGRESQL,
-        [SEMATTRS_DB_NAME]: CONFIG.database,
-        [SEMATTRS_NET_PEER_NAME]: CONFIG.host,
-        [SEMATTRS_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
-        [SEMATTRS_NET_PEER_PORT]: CONFIG.port,
-        [SEMATTRS_DB_USER]: CONFIG.user,
+        [ATTR_DB_SYSTEM]: DB_SYSTEM_VALUE_POSTGRESQL,
+        [ATTR_DB_NAME]: CONFIG.database,
+        [ATTR_NET_PEER_NAME]: CONFIG.host,
+        [ATTR_DB_CONNECTION_STRING]: `postgresql://${CONFIG.host}:${CONFIG.port}/${CONFIG.database}`,
+        [ATTR_NET_PEER_PORT]: CONFIG.port,
+        [ATTR_DB_USER]: CONFIG.user,
         [AttributeNames.IDLE_TIMEOUT_MILLIS]: CONFIG.idleTimeoutMillis,
       };
 
@@ -258,7 +263,7 @@ describe('pg-pool', () => {
       };
       const pgAttributes = {
         ...DEFAULT_PG_ATTRIBUTES,
-        [SEMATTRS_DB_STATEMENT]: 'SELECT NOW()',
+        [ATTR_DB_STATEMENT]: 'SELECT NOW()',
       };
       const events: TimedEvent[] = [];
       const parentSpan = provider
@@ -332,7 +337,7 @@ describe('pg-pool', () => {
       };
       const pgAttributes = {
         ...DEFAULT_PG_ATTRIBUTES,
-        [SEMATTRS_DB_STATEMENT]: 'SELECT NOW()',
+        [ATTR_DB_STATEMENT]: 'SELECT NOW()',
       };
       const events: TimedEvent[] = [];
       const span = provider.getTracer('test-pg-pool').startSpan('test span');
@@ -351,7 +356,7 @@ describe('pg-pool', () => {
       };
       const pgAttributes = {
         ...DEFAULT_PG_ATTRIBUTES,
-        [SEMATTRS_DB_STATEMENT]: 'SELECT NOW()',
+        [ATTR_DB_STATEMENT]: 'SELECT NOW()',
       };
       const events: TimedEvent[] = [];
       const parentSpan = provider
@@ -388,7 +393,7 @@ describe('pg-pool', () => {
         };
         const pgAttributes = {
           ...DEFAULT_PG_ATTRIBUTES,
-          [SEMATTRS_DB_STATEMENT]: query,
+          [ATTR_DB_STATEMENT]: query,
           [dataAttributeName]: '{"rowCount":1}',
         };
 
@@ -470,7 +475,7 @@ describe('pg-pool', () => {
         };
         const pgAttributes = {
           ...DEFAULT_PG_ATTRIBUTES,
-          [SEMATTRS_DB_STATEMENT]: query,
+          [ATTR_DB_STATEMENT]: query,
         };
 
         beforeEach(async () => {
@@ -967,5 +972,120 @@ describe('pg-pool (ESM)', () => {
         }
       },
     });
+  });
+});
+
+describe('pg semantic conventions env variable', () => {
+  let pool: pgPool<pg.Client>;
+  let contextManager: AsyncLocalStorageContextManager;
+  let instrumentation: PgInstrumentation;
+  const provider = new BasicTracerProvider({
+    spanProcessors: [new SimpleSpanProcessor(memoryExporter)],
+  });
+  const testPostgres = process.env.RUN_POSTGRES_TESTS;
+  const shouldTest = testPostgres;
+  // Here we are `require`ing *before* the instrumentation is created below.
+  // In *general* this is a potential instrumentation issue, but this works for
+  // `pg-pool` instrumentation because it patches the `pgPool.prototype`
+  // (i.e. not a top-level export).
+  const pgPool = require('pg-pool');
+  pool = new pgPool(CONFIG);
+
+  before(function () {
+    const skip = () => {
+      // this.skip() workaround
+      // https://github.com/mochajs/mocha/issues/2683#issuecomment-375629901
+      this.test!.parent!.pending = true;
+      this.skip();
+    };
+
+    if (!shouldTest) {
+      skip();
+    }
+  });
+
+  beforeEach(() => {
+    contextManager = new AsyncLocalStorageContextManager().enable();
+    context.setGlobalContextManager(contextManager);
+  });
+
+  afterEach(() => {
+    delete process.env.OTEL_SEMCONV_STABILITY_OPT_IN;
+    memoryExporter.reset();
+    context.disable();
+  });
+
+  it('send only default semantic conventions', async () => {
+    process.env.OTEL_SEMCONV_STABILITY_OPT_IN = '';
+    instrumentation = new PgInstrumentation();
+    instrumentation.setTracerProvider(provider);
+    pool = new pgPool(CONFIG);
+
+    const pgAttributes = {
+      ...DEFAULT_PG_ATTRIBUTES,
+      [ATTR_DB_STATEMENT]: 'SELECT NOW()',
+    };
+    const events: TimedEvent[] = [];
+    const span = provider
+      .getTracer('test-pg-pool-semantic')
+      .startSpan('test span');
+    await context.with(trace.setSpan(context.active(), span), async () => {
+      const result = await pool.query('SELECT NOW()');
+      runCallbackTest(span, pgAttributes, events, unsetStatus, 3, 2);
+      assert.ok(result, 'pool.query() returns a promise');
+    });
+
+    span.end();
+    await pool.end();
+  });
+
+  it('send both default and stable semantic conventions', async () => {
+    process.env.OTEL_SEMCONV_STABILITY_OPT_IN = 'database/dup';
+    instrumentation = new PgInstrumentation();
+    instrumentation.setTracerProvider(provider);
+    pool = new pgPool(CONFIG);
+
+    const pgAttributes = {
+      ...DEFAULT_PG_ATTRIBUTES,
+      ...STABLE_PG_QUERY_ATTRIBUTES,
+      [ATTR_DB_STATEMENT]: 'SELECT NOW()',
+      [ATTR_DB_QUERY_TEXT]: 'SELECT NOW()',
+    };
+    const events: TimedEvent[] = [];
+    const span = provider
+      .getTracer('test-pg-pool-semantic')
+      .startSpan('test span');
+    await context.with(trace.setSpan(context.active(), span), async () => {
+      const result = await pool.query('SELECT NOW()');
+      runCallbackTest(span, pgAttributes, events, unsetStatus, 3, 2);
+      assert.ok(result, 'pool.query() returns a promise');
+    });
+
+    span.end();
+    await pool.end();
+  });
+
+  it('send only stable semantic conventions', async () => {
+    process.env.OTEL_SEMCONV_STABILITY_OPT_IN = 'database';
+    instrumentation = new PgInstrumentation();
+    instrumentation.setTracerProvider(provider);
+    pool = new pgPool(CONFIG);
+
+    const pgAttributes = {
+      ...STABLE_PG_QUERY_ATTRIBUTES,
+      [ATTR_DB_QUERY_TEXT]: 'SELECT NOW()',
+    };
+    const events: TimedEvent[] = [];
+    const span = provider
+      .getTracer('test-pg-pool-semantic')
+      .startSpan('test span');
+    await context.with(trace.setSpan(context.active(), span), async () => {
+      const result = await pool.query('SELECT NOW()');
+      runCallbackTest(span, pgAttributes, events, unsetStatus, 3, 2);
+      assert.ok(result, 'pool.query() returns a promise');
+    });
+
+    span.end();
+    await pool.end();
   });
 });